refactor(modal): extract ModalProvider and ModalActions to separate files
